added real html/tsx forms to dev server
