+++ ui-packages/packages/runtime-tools-dev-ui-webapp/server/MockData/forms/examples/html_hiring_HRInterview.config
@@ -1,0 +1,13 @@
+{
+    "resources": {
+        "styles": {
+            "bootstrap.min.css": "https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/css/bootstrap.min.css"
+        },
+        "scripts": {
+            "bootstrap.min.js": "https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/js/bootstrap.min.js",
+            "jquery.js": "https://code.jquery.com/jquery-3.2.1.slim.min.js",
+            "popper.js": "https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.12.9/umd/popper.min.js"
+        }
+    },
+    "schema": "{\"$schema\":\"http://json-schema.org/draft-07/schema#\",\"type\":\"object\",\"properties\":{\"candidate\":{\"type\":\"object\",\"properties\":{\"email\":{\"type\":\"string\"},\"name\":{\"type\":\"string\"},\"salary\":{\"type\":\"integer\"},\"skills\":{\"type\":\"string\"}},\"input\":true},\"approve\":{\"type\":\"boolean\",\"output\":true}}}"
+}
