+++ ui-packages/packages/runtime-tools-dev-ui-webapp/server/MockData/forms/examples/html_hiring_HRInterview.html
@@ -1,0 +1,31 @@
+<div>
+  <fieldset disabled>
+    <legend>Candidate</legend>
+    <div>
+      <div class="form-group">
+        <label for="uniforms-0000-0002">Email</label>
+        <input type="text" id="uniforms-0000-0002" name="candidate.email" class="form-control" disabled value="" />
+      </div>
+
+      <div class="form-group">
+        <label for="uniforms-0000-0003">Name</label>
+        <input type="text" id="uniforms-0000-0003" name="candidate.name" class="form-control" disabled value="" />
+      </div>
+
+      <div class="form-group">
+        <label for="uniforms-0000-0005">Salary</label>
+        <input type="number" class="form-control" id="uniforms-0000-0005" name="candidate.salary" disabled step="1" value="" />
+      </div>
+
+      <div class="form-group">
+        <label for="uniforms-0000-0006">Skills</label>
+        <input type="text" id="uniforms-0000-0006" name="candidate.skills" class="form-control" disabled value="" />
+      </div>
+    </div>
+  </fieldset>
+
+  <div class="form-check">
+    <input type="checkbox" id="uniforms-0000-0008" name="approve" class="form-check-input" />
+    <label class="form-check-label" for="uniforms-0000-0008">Approve</label>
+  </div>
+</div>
