+++ ui-packages/packages/runtime-tools-dev-ui-webapp/server/MockData/forms/examples/html_hiring_ITInterview.html
@@ -1,0 +1,31 @@
+<div>
+  <div class="form-check">
+    <input type="checkbox" id="uniforms-0001-0001" name="approve" class="form-check-input" />
+    <label class="form-check-label" for="uniforms-0001-0001">Approve</label>
+  </div>
+
+  <fieldset disabled>
+    <legend>Candidate</legend>
+    <div>
+      <div class="form-group">
+        <label for="uniforms-0001-0004">Email</label>
+        <input type="text" id="uniforms-0001-0004" name="candidate.email" class="form-control" disabled value="" />
+      </div>
+
+      <div class="form-group">
+        <label for="uniforms-0001-0005">Name</label>
+        <input type="text" id="uniforms-0001-0005" name="candidate.name" class="form-control" disabled value="" />
+      </div>
+
+      <div class="form-group">
+        <label for="uniforms-0001-0007">Salary</label>
+        <input type="number" class="form-control" id="uniforms-0001-0007" name="candidate.salary" disabled step="1" value="" />
+      </div>
+
+      <div class="form-group">
+        <label for="uniforms-0001-0008">Skills</label>
+        <input type="text" id="uniforms-0001-0008" name="candidate.skills" class="form-control" disabled value="" />
+      </div>
+    </div>
+  </fieldset>
+</div>
