+++ ui-packages/packages/runtime-tools-dev-ui-webapp/server/MockData/forms/examples/react_hiring_HRInterview.config
@@ -1,0 +1,7 @@
+{
+    "resources": {
+        "styles": {},
+        "scripts": {}
+    },
+    "schema": "{\"$schema\":\"http://json-schema.org/draft-07/schema#\",\"type\":\"object\",\"properties\":{\"candidate\":{\"type\":\"object\",\"properties\":{\"email\":{\"type\":\"string\"},\"name\":{\"type\":\"string\"},\"salary\":{\"type\":\"integer\"},\"skills\":{\"type\":\"string\"}},\"input\":true},\"approve\":{\"type\":\"boolean\",\"output\":true}}}"
+}
