+++ ui-packages/packages/runtime-tools-dev-ui-webapp/server/MockData/forms/formData.js
@@ -1,21 +1,21 @@
 module.exports = [
     {
-        name: 'html_HRInterview',
+        name: 'html_hiring_HRInterview',
         type: 'HTML',
         lastModified: new Date('2021-08-23T13:26:02.13Z')
     },
     {
-        name: 'html_ITInterview',
+        name: 'html_hiring_ITInterview',
         type: 'HTML',
         lastModified: new Date('2021-08-23T13:26:02.13Z')
     },
     {
-        name: 'hiring_HRInterview',
+        name: 'react_hiring_HRInterview',
         type: 'TSX',
         lastModified: new Date('2021-08-23T13:26:02.13Z')
     },
     {
-        name: 'hiring_ITInterview',
+        name: 'react_hiring_ITInterview',
         type: 'TSX',
         lastModified: new Date('2021-08-23T13:26:02.13Z')
     },
